refactor: ディレクトリ構造を整理
- refactor: `SearchResultMessageFileList*` を `MessageFileSummary*` に rename
- refactor: `MessageFileList*`, `MessageFileSummary*`, `MessageQuoteList*`, `MessageOgpList*` を `MessageElement/Embeddings` 下へ移動
