+++ src/components/Main/MainView/MessageElement/Embeddings/MessageFileListItemContent.vue
