+++ src/components/Main/MainView/MessageElement/Embeddings/MessageFileSummaryImage.vue
