+++ src/components/Main/MainView/MessageElement/Embeddings/MessageOgpDescription.vue
