+++ src/components/Main/MainView/MessageElement/Embeddings/MessageQuoteList.vue
