+++ src/components/Main/MainView/MessageElement/Embeddings/MessageQuoteListItemFooter.vue
