+++ src/components/Main/MainView/MessageElement/Embeddings/MessageQuoteListItemHeader.vue
